% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averages.R
\name{calculate_blood_type_averages}
\alias{calculate_blood_type_averages}
\title{Calculate the average of each blood type, including Country and Population}
\usage{
calculate_blood_type_averages(filepath)
}
\arguments{
\item{filepath}{Path to the CSV file containing blood type data.}
}
\value{
A tibble with BloodType and AverageFrequency columns sorted by descending AverageFrequency.
}
\description{
This function calculates the average frequency for each blood type.
}
\examples{
calculate_blood_type_averages("/path/to/your/file.csv")
}
