% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stat.R
\name{calculate_blood_type_summary_stats}
\alias{calculate_blood_type_summary_stats}
\title{Calculate summary statistics for each blood type}
\usage{
calculate_blood_type_summary_stats(filepath)
}
\arguments{
\item{filepath}{Path to the CSV file containing blood type data.}
}
\description{
This function calculates summary statistics (e.g., mean, median, etc.) for each blood type.
}
\examples{
calculate_blood_type_summary_stats("/path/to/your/file.csv")

}
