% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{create_blood_type_bar_plot}
\alias{create_blood_type_bar_plot}
\title{Create bar plots for blood type frequencies by country}
\usage{
create_blood_type_bar_plot(data)
}
\arguments{
\item{data}{A tibble containing blood type data with columns for each blood type.}
}
\value{
A ggplot object.
}
\description{
This function creates bar plots for blood type frequencies by country.
}
\examples{
create_blood_type_bar_plot(blood_type_data)

}
