% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{create_blood_type_boxplots}
\alias{create_blood_type_boxplots}
\title{Create box plots for each blood type}
\usage{
create_blood_type_boxplots(data)
}
\arguments{
\item{data}{A tibble containing blood type data with columns for each blood type.}
}
\value{
A ggplot object.
}
\description{
This function creates box plots for each blood type.
}
\examples{
create_blood_type_boxplots(blood_type_data)

}
