% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{create_blood_type_histogram}
\alias{create_blood_type_histogram}
\title{Create histograms for each blood type with colors}
\usage{
create_blood_type_histogram(blood_type_data)
}
\arguments{
\item{blood_type_data}{A tibble containing blood type data with columns for each blood type.}
}
\value{
A ggplot object.
}
\description{
This function creates histograms for each blood type with different colors.
}
\examples{
create_blood_type_histogram(blood_type_data)

}
